+++ toucan-shopping-modules/product/toucan-shopping-product-entity-vo/src/main/java/com/toucan/shopping/modules/product/entity/ProductSpuAttributeValue.java
@@ -33,7 +33,7 @@ public class ProductSpuAttributeValue {
     private Long attributeValueId;
 
     private String attributeValue; //属性值
-    private Integer attributeSort; //排序
+    private Long attributeSort; //排序
     private String attributeValueExtend1; //属性值扩展字段
     private String attributeValueExtend2; //属性值扩展字段
     private String attributeValueExtend3; //属性值扩展字段
